requirements.txt
discord.py>=2.3,<3.0
python-dotenv>=1.0,<2.0
anthropic>=0.42,<1.0
scikit-learn>=1.3,<2.0
